backend/internal_medicine/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
httpx>=0.27
numpy==1.26.4
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
transformers==4.44.2
accelerate>=0.33.0
tqdm>=4.66
ijson>=3.2
# Optional (needed for some HF seq2seq models like Flan-T5):
sentencepiece>=0.2.0
protobuf>=4.25.0
# Optional (only needed for SciSpacy sentence splitting + BC5CDR NER):
# spacy==3.7.5
# scispacy==0.5.4
# Then install models:
#   pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz
#   pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz
